+++ java-io-http/src/main/java/module-info.java
@@ -5,6 +5,7 @@ module nbbrd.io.http {
     requires static nbbrd.design;
 
     requires nbbrd.io.base;
+    requires static okhttp3;
 
     exports nbbrd.io.http;
     exports nbbrd.io.http.ext;